refactor: clean cache retrieval code
